added controller for user
